Implemented a "full savestate-movie load" mode similar to the implementation in VBA-rr & SNES9x-rr.
In this mode loading a savestate in read+write doesn't truncate the movie to its frame count immediately.  Instead it waits 

until input is recording into the movie (next frame).  For win32 this feature is togglable in movie options and the context 

menu.  For SDL this is off by default and a toggle will need to be added.